Refactor CalculatedColumnModal to improve maintainability
Extracted the formula editing logic (autocomplete, bracket pairing,
suggestions) into a separate hook, `useFormulaEditor.ts`. Extracted
the large list of available shortcuts into a child component,
`FormulaShortcuts.tsx`. Updated `CalculatedColumnModal.tsx` to use
the shared `Modal` component from `src/components/Layout/Modal.tsx`
instead of recreating the modal overlay and headers manually.
